Revert "updated content written to data.json"
This reverts commit 22a7ae9cd310854e29ffd3abe1be7b8c7d8cbf65.

When you submit an app to App Store Connect, you have to include all of the targets. This means that if you have a WatchOS App that isn't ready, you'll have to submit it when uploading the app. I decided to simply delete it on my lcoal machine and revert to this commit which is the same, except that it has the WatchOS target intact.